Separating out blocks from models.py
cc @ccostino
